requirements.txt
# InvoiceGemini - AI-Powered Invoice Data Extraction
# Production Dependencies

# =============================================================================
# Core GUI Framework
# =============================================================================
PyQt6>=6.9.0
PyQt6-sip>=13.10.0
PyQt6-Qt6>=6.9.0

# =============================================================================
# Machine Learning and AI Models
# =============================================================================
# PyTorch (CPU version - users can install CUDA version if needed)
torch>=2.0.0
torchvision>=0.15.0

# Hugging Face Transformers
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.14.0
sentencepiece>=0.1.99
accelerate>=0.20.0
evaluate>=0.4.0
seqeval>=1.2.2

# Google Gemini API
google-generativeai>=0.8.0

# =============================================================================
# Computer Vision and OCR
# =============================================================================
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
PyMuPDF>=1.23.0

# =============================================================================
# Data Processing and Analysis
# =============================================================================
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# =============================================================================
# Export and File Handling
# =============================================================================
openpyxl>=3.1.0
beautifulsoup4>=4.12.0

# =============================================================================
# Configuration and Environment
# =============================================================================
python-dotenv>=1.0.0
cryptography>=41.0.0

# =============================================================================
# API Integration
# =============================================================================
requests>=2.31.0

# =============================================================================
# System Monitoring and Performance
# =============================================================================
psutil>=5.9.0
tqdm>=4.65.0

# =============================================================================
# Development and Testing
# =============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# =============================================================================
# Optional GPU Support
# =============================================================================
# For CUDA support, install manually:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# 
# Additional GPU monitoring (optional):
# nvidia-ml-py>=12.535.133
# GPUtil>=0.4.0 